% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_brain.R
\name{clean_brain}
\alias{clean_brain}
\title{Takes a emocemp brain mri .csv file and clean it}
\usage{
clean_brain(datafile)
}
\arguments{
\item{datafile}{File extracted from qualtrics in .csv format with the numeric option}
}
\value{
a data frame with the input brain mri data cleaned
}
\description{
Takes a emocemp brain mri .csv file and clean it
}
\author{
R.C.S
}
