% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_orbit.R
\name{clean_orbit}
\alias{clean_orbit}
\title{Takes a emocemp orbit mri .csv file and clean it}
\usage{
clean_orbit(datafile)
}
\arguments{
\item{datafile}{File extracted from qualtrics in .csv format with the numeric option}
}
\value{
a data frame with the input brain mri data cleaned
}
\description{
Takes a emocemp orbit mri .csv file and clean it
}
\author{
R.C.S
}
