% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_spinal.R
\name{clean_spinal}
\alias{clean_spinal}
\title{Takes a EMOCEMP spinal cord MRI .csv (numeric) file and clean it}
\usage{
clean_spinal(datafile)
}
\arguments{
\item{datafile}{File extracted from Qualtrics in .csv format with the numeric option}
}
\value{
a data frame with the input brain MRI data cleaned
}
\description{
Takes a EMOCEMP spinal cord MRI .csv (numeric) file and clean it
}
\author{
R.C.S
}
