% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tele.R
\name{clean_tele}
\alias{clean_tele}
\title{Takes a raw dataset for emocemp teleconsults and clean it}
\usage{
clean_tele(datafile)
}
\arguments{
\item{datafile}{a EMOCEMP teleconsult .csv file exported from qualtrics with numeric data option}
}
\value{
a data.frame with cleaned data which may be used to merge with other visits
}
\description{
The raw data need to be exported from qualtrics EMOCEMP Tele in .csv with the numeric values option checked
}
\author{
R.C.S
}
