% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_v2.R
\name{clean_v2}
\alias{clean_v2}
\title{Takes the raw dataset from qualtrics and clean it}
\usage{
clean_v2(datafile)
}
\arguments{
\item{datafile}{a V2 .csv file exported from qualtrics with numeric data option}
}
\value{
a data.frame with cleaned data which may be used to merge with other visits
}
\description{
The raw data need to be exported from qualtrics EMOCEMP V2 in .csv with the numeric values option checked
}
\author{
R.C.S
}
