% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohen_kappa_mri.R
\name{cohen_kappa_mri}
\alias{cohen_kappa_mri}
\title{Takes a cleaned data frame from MRI study and calculate cohen kappa for selected varibles}
\usage{
cohen_kappa_mri(datafile, varnames)
}
\arguments{
\item{datafile}{a data frame resulted from clean_brain or clean_orbit}

\item{varnames}{a string vector with variable names from the datasets}
}
\value{
a list with kappa statistics for each selected variable
}
\description{
Takes a cleaned data frame from MRI study and calculate cohen kappa for selected varibles
}
