% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummie_mcv.R
\name{dummie_mcv}
\alias{dummie_mcv}
\title{Combines mutate_mcv and split_mcv to clean datasets}
\usage{
dummie_mcv(names, var, n, dataf)
}
\arguments{
\item{names}{a vector of strings with names in the order of correspondent numeric values}

\item{var}{a variable from the database with multiple choice answers}

\item{n}{string containing the number of possible answers for the selected var}

\item{dataf}{a data frame with numeric values associated with the desired factor}
}
\value{
New data frame with the values distributed across subjects in the approriate factor level
}
\description{
Combines mutate_mcv and split_mcv to clean datasets
}
\author{
R.C.S
}
