% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_mcv.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Takes a variable with multiple numeric variables and split into a data.frame with unique values in each col}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{var}{Takes a string variable from a data frame with 'n' possible values}

\item{n}{Integer with the maximum number of possible values in a individual subject}

\item{dataf}{Data frame from which the variable will be taken}
}
\value{
a data frame
}
\description{
Takes a variable with multiple numeric variables and split into a data.frame with unique values in each col
}
\author{
R.C.S
}
