% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_v1_v2.R
\name{merge_v1v2}
\alias{merge_v1v2}
\title{Merge visit 1 and visit 2 data frames}
\usage{
merge_v1v2(v1, v2, mode = c("long", "wide"), all.patients = FALSE)
}
\arguments{
\item{v1}{data frame resulted from \link{clean_v1}}

\item{v2}{data frame resulted from \link{clean_v2}}

\item{mode}{select "long" or "wide" for different merge options. If a single line for each subject is desired = "wide". If a line for each visit is desired ; = "long"}

\item{all.patients}{Default = FALSE. If TRUE is selected, all patients with visit 1 will be merged, and NA values will be inputted in specific visit 2 variables}
}
\value{
a data frame with merged visit 1 and 2 as specified in the parameters
}
\description{
Merge visit 1 and visit 2 data frames
}
\author{
R.C.S
}
