% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_visits.R
\name{merge_visits}
\alias{merge_visits}
\title{Merge data from different visits. Set at least two cleaned visit data files.}
\usage{
merge_visits(v1, v2, v3 = NULL, mode = c("long", "wide"), all.patients = FALSE)
}
\arguments{
\item{v1}{data frame resulted from \link{clean_v1}}

\item{v2}{data frame resulted from \link{clean_v2}}

\item{v3}{data frame resulted from \link{clean_v3} ; Default is = NULL}

\item{mode}{select "long" or "wide" for different merge options. If a single line for each subject is desired = "wide". If a line for each visit is desired ; = "long"}

\item{all.patients}{Default = FALSE. If TRUE is selected, all patients with visit 1 will be merged, and NA values will be inputted in specific visit 2 variables}
}
\value{
a data frame with merged visit 1 and 2 as specified in the parameters
}
\description{
Merge data from different visits. Set at least two cleaned visit data files.
}
\author{
R.C.S
}
