% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_mcv.R
\name{mutate_mcv}
\alias{mutate_mcv}
\title{Creates a new data.frame (new_df) with appropriated variable names and values (dummies)}
\usage{
mutate_mcv(names, df)
}
\arguments{
\item{names}{a vector of strings in the order of correspondent numeric values}

\item{df}{a data frame with numeric values associated with the desired factor levels}
}
\value{
New data frame with the values distributed across subjects in the approriate factor level
}
\description{
Takes a vector of strings and a data frame, usually resulted from split_mcv()
}
\author{
R.C.S
}
