% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_mcv.R
\name{split_mcv}
\alias{split_mcv}
\title{Takes a variable with multiple numeric variables and split into a data.frame with unique values in each col}
\usage{
split_mcv(var, n, dataf)
}
\arguments{
\item{var}{Takes a string variable from a data frame with 'n' possible values}

\item{n}{Integer with the maximum number of possible values in a individual subject}

\item{dataf}{Data frame from which the variable will be taken}
}
\value{
a data frame
}
\description{
Takes a variable with multiple numeric variables and split into a data.frame with unique values in each col
}
\author{
R.C.S
}
