% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_v1.R
\name{clean_v1}
\alias{clean_v1}
\title{Takes the raw .csv file exported from qualtrics and clean it}
\usage{
clean_v1(datafile)
}
\arguments{
\item{datafile}{a V1 .csv file exported from qualtrics with numeric data option}
}
\description{
Takes the raw .csv file exported from qualtrics and clean it
}
\author{
R.C.S
}
