% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Test if a numeric input is an integer}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric value to test}

\item{tol}{limit defined as a really low number}
}
\description{
Test if a numeric input is an integer
}
